LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/offset-anchor-expected.txt

CONSOLE MESSAGE: Error: '<position>' is not a valid CSS component

Harness Error (FAIL), message = Error: '<position>' is not a valid CSS component

PASS Can set 'offset-anchor' to CSS-wide keywords
FAIL Can set 'offset-anchor' to var() references assert_equals: expected 2 but got 1
PASS Can set 'offset-anchor' to the 'auto' keyword
